Release v1.9.8
* Optimize plugins, removal of unnecessary function calls

+++ plugins/modules/iam_account_alias.py
@@ -59,15 +59,11 @@ state:
   type: str
 """
 
-from ansible.module_utils.common.dict_transformations import snake_dict_to_camel_dict
 from ansible_collections.amazon.aws.plugins.module_utils.botocore import (
     paginated_query_with_retries,
 )
 from ansible_collections.amazon.aws.plugins.module_utils.modules import AnsibleAWSModule
 from ansible_collections.amazon.aws.plugins.module_utils.retries import AWSRetry
-from ansible_collections.amazon.aws.plugins.module_utils.transformation import (
-    scrub_none_parameters,
-)
 
 
 def ensure_absent(client, module):
@@ -83,11 +79,7 @@ def ensure_absent(client, module):
     if changed and not module.check_mode:
         try:
             client.delete_account_alias(
-                **scrub_none_parameters(
-                    snake_dict_to_camel_dict(
-                        {"account_alias": name}, capitalize_first=True
-                    )
-                ),
+                AccountAlias=name,
                 aws_retry=True,
             )
         except Exception as e:
@@ -117,14 +109,12 @@ def ensure_present(client, module):
     changed = aliases != desired_aliases
 
     if changed and not module.check_mode:
-        for alias in [alias for alias in aliases if alias not in desired_aliases]:
+        for alias in aliases:
+            if alias in desired_aliases:
+                continue
             try:
                 client.delete_account_alias(
-                    **scrub_none_parameters(
-                        snake_dict_to_camel_dict(
-                            {"account_alias": alias}, capitalize_first=True
-                        )
-                    ),
+                    AccountAlias=alias,
                     aws_retry=True,
                 )
             except Exception as e:
@@ -132,14 +122,12 @@ def ensure_present(client, module):
                     e, msg=f"Unable to delete AWS IAM account alias {alias}"
                 )
 
-        for alias in [alias for alias in desired_aliases if alias not in aliases]:
+        for alias in desired_aliases:
+            if alias in aliases:
+                continue
             try:
                 client.create_account_alias(
-                    **scrub_none_parameters(
-                        snake_dict_to_camel_dict(
-                            {"account_alias": alias}, capitalize_first=True
-                        )
-                    ),
+                    AccountAlias=alias,
                     aws_retry=True,
                 )
             except Exception as e:
